+++ src/Starlight/Radar.hs
@@ -14,9 +14,9 @@ import Unit.Power
 -- FIXME: transmitter gain
 -- FIXME: effective aperture
 newtype Radar = Radar
-  { power :: Mega Watts Float
+  { power :: Mega Watts Double
   }
   deriving (Generic, Show)
 
-power_ :: Lens' Radar (Mega Watts Float)
+power_ :: Lens' Radar (Mega Watts Double)
 power_ = field @"power"
